requirements.txt
annotated-types==0.7.0
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
marshmallow==3.21.3
more-itertools==10.4.0
mypy==1.11.1
mypy-extensions==1.0.0
Nasdaq-Data-Link==1.0.4
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pandas-stubs==2.2.2.240603
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pytest==8.3.2
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.32
types-colorama==0.4.15.20240311
types-docutils==0.21.0.20240724
types-psycopg2==2.9.21.20240417
types-Pygments==2.18.0.20240506
types-pytz==2024.1.0.20240417
types-requests==2.32.0.20240712
types-setuptools==71.1.0.20240726
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
